# gemip-mipcommonapi/src/main/java/it/csi/gemip/mipcommonapi/api/ExcelApi.java
package it.csi.gemip.mipcommonapi.api;

/*-
 * Copyright © 2025 Regione Piemonte
 *
 * Licensed under the EUPL, Version 1.2 or – as soon they will be
 * approved by the European Commission - subsequent versions of the
 * EUPL (the "Licence");
 *
 * You may not use this work except in compliance with the Licence.
 * You may obtain a copy of the Licence at:
 *
 * https://interoperable-europe.ec.europa.eu/collection/eupl/eupl-text-eupl-12
 * or
 * https://opensource.org/license/EUPL-1.2
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the Licence is distributed on an "AS IS" basis,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the Licence for the specific language governing permissions and
 * limitations under the Licence.
 */

import it.csi.gemip.mipcommonapi.api.dto.*;


import it.csi.gemip.mipcommonapi.api.dto.Errore;
import java.io.File;

import java.util.List;
import java.util.Map;

import java.io.InputStream;

import jakarta.ws.rs.core.Context;
import jakarta.ws.rs.core.Response;
import jakarta.ws.rs.core.SecurityContext;
import jakarta.ws.rs.core.HttpHeaders;
import jakarta.servlet.http.HttpServletRequest;
import jakarta.ws.rs.*;
import jakarta.validation.constraints.*;
@Path("/excel")


public interface ExcelApi  {
   
    /**
     * summary = esportazione del foglio excel
     * description = restituisce il file excel degli anagrafica cittadino
     * @return Response
     * responses: 
       <ul>
         <li>    
           <p>responseCode = 200, description = excel<br>
           schema implementation = { @see File }</p>
         </li>
         <li>    
           <p>responseCode = 400, description = usata quando i campi non sono validati correttamente<br>
           schema implementation = { @see Errore }</p>
         </li>
         <li>    
           <p>responseCode = 403, description = utente non autorizzato al caso d'uso<br>
           schema implementation = { @see Errore }</p>
         </li>
         <li>    
           <p>responseCode = 200, description = errore generico<br>
           schema implementation = { @see Errore }</p>
         </li>
       </ul>
    */
    @GET
    @Path("/esporta-excel-anagrafica-cittadino")
    
    @Produces({ "application/octet-stream", "application/json" })
    Response exportExcelAnagraficaCittadino(@NotNull


@jakarta.validation.Valid  @QueryParam("idIncontroPreaccoglienza") Long idIncontroPreaccoglienza,@NotNull


@jakarta.validation.Valid  @QueryParam("codOperatore") String codOperatore,


@jakarta.validation.Valid  @QueryParam("idSoggettoAttuatore") String idSoggettoAttuatore,@Context SecurityContext securityContext, @Context HttpHeaders httpHeaders , @Context HttpServletRequest httpRequest );

    /**
     * summary = esportazione del foglio excel contenete i partecipanti dgli incontri
     * description = restituisce la lista degli incontri di preaccoglienza con il pritrto su tutti i cittadini
     * @return Response
     * responses: 
       <ul>
         <li>    
           <p>responseCode = 200, description = excel<br>
           schema implementation = { @see File }</p>
         </li>
         <li>    
           <p>responseCode = 400, description = usata quando i campi non sono validati correttamente<br>
           schema implementation = { @see Errore }</p>
         </li>
         <li>    
           <p>responseCode = 403, description = utente non autorizzato al caso d'uso<br>
           schema implementation = { @see Errore }</p>
         </li>
         <li>    
           <p>responseCode = 200, description = errore generico<br>
           schema implementation = { @see Errore }</p>
         </li>
       </ul>
    */
    @GET
    @Path("/export-preaccoglienza-anagrafica")
    
    @Produces({ "application/octet-stream", "application/json" })
    Response getExportIncontriPreaccoglienzaAnagrafica(


@jakarta.validation.Valid  @QueryParam("idIncontriPreaccoglienza") String idIncontriPreaccoglienza,


@jakarta.validation.Valid  @QueryParam("codOperatore") String codOperatore,


@jakarta.validation.Valid  @QueryParam("idSoggettoAttuatore") String idSoggettoAttuatore,@Context SecurityContext securityContext, @Context HttpHeaders httpHeaders , @Context HttpServletRequest httpRequest );

}