# gemip-mipcommonapi/src/main/resources/META-INF/openapi.yaml
openapi: 3.0.1
info:
  title: API mipcommonapi
  description: API comuni mipcittadino e mipbackoffice
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local dev environment
paths:
...
  /comuni:
    get:
      summary: elenco comuni
      tags:
        - dati comuni
      operationId: getComuni
      description: elenco dei comuni
      responses:
        '200':
          description: elenco dei comuni
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comune"
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: errore generico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errore'
  /comuni/provincia:
    get:
      summary: elenco comuni della provincia
      tags:
        - dati comuni
      operationId: getComuniPerProvincia
      description: elenco delle comuni
      parameters:
        - in: query
          name: provincia
          schema:
            type: string
          required: true
          description: codice della provincia a cercare
      responses:
        '200':
          description: elenco dei comuni associate ad una provincia
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comune"
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: errore generico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errore'
...
  /excel/esporta-excel-anagrafica-cittadino:
    get:
      summary: esportazione del foglio excel
      tags:
        - excel
      operationId: exportExcelAnagraficaCittadino
      description: restituisce il file excel degli anagrafica cittadino
      parameters:
        - in: query
          name: idIncontroPreaccoglienza
          schema:
            type: integer
            format: int64
          required: true
        - in: query
          name: codOperatore
          schema:
            type: string
          required: true
        - in: query
          name: idSoggettoAttuatore
          schema:
            type: string
          required: false
      responses:
        '200':
          description: excel
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: errore generico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errore'
  /excel/export-preaccoglienza-anagrafica:
    get:
      summary: esportazione del foglio excel contenete i partecipanti dgli incontri
      tags:
        - excel
      operationId: getExportIncontriPreaccoglienzaAnagrafica
      description: restituisce la lista degli incontri di preaccoglienza con il pritrto su tutti i cittadini
      parameters:
        - in: query
          name: idIncontriPreaccoglienza
          schema:
            type: string
          required: false
          description: codice area territoriale
        - in: query
          name: codOperatore
          schema:
            type: string
          required: false
          description: codice operatore
        - in: query
          name: idSoggettoAttuatore
          schema:
            type: string
          required: false
          description: id soggetto attuatore
      responses:
        '200':
          description: excel
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/InvalidParameter'
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          description: errore generico
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Errore'
...
components:
  schemas:
...
    Comune:
      type: object
      properties:
        codiceIstatComune:
          type: string
        provincia:
          $ref: "#/components/schemas/Provincia"
        descrizioneComune:
          type: string
        cap:
          type: string
        codiceFiscaleComune:
          type: string
...
    Provincia:
      type: object
      properties:
        codiceProvincia:
          type: string
        codiceRegione:
          type: string
        descrizioneProvincia:
          type: string
        siglaProvincia:
          type: string
...
    Errore:
      type: object
      properties:
        title:
          type: string
        status:
          type: integer
        code:
          type: integer
        messaggioCittadino:
          type: string
        errorMessage:
          type: string
        fields:
          type: string
          nullable: true
  responses:
    InvalidParameter:
      description: usata quando i campi non sono validati correttamente
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'
    Forbidden:
      description: utente non autorizzato al caso d'uso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Errore'